CI Test/requirements.txt

langchain
langchain-community
langchain-google-genai
langchain-openai
langchain-together

numpy
pandas

pillow
pydantic
pydantic-settings
python-dotenv
python-multipart
sentence-transformers
starlette
tiktoken
uvicorn
